Move diff_report.py code into codediff.py and add subcommands
