+++ tools/codediff/requirements.txt
@@ -1,1 +1,2 @@
+dataclasses-json
 jinja2
